[Interfaces] Split SPI and I2C config into new Interfaces page
