+++ src/src/WebServer/HardwarePage.h
@@ -10,12 +10,6 @@
 // ********************************************************************************
 void handle_hardware();
 
-# if FEATURE_PLUGIN_PRIORITY
-bool isI2CPriorityTaskActive(uint8_t i2cBus);
-void I2CShowSdaSclReadonly(int8_t  i2c_sda,
-                           int8_t  i2c_scl,
-                           uint8_t i2cBus);
-# endif // if FEATURE_PLUGIN_PRIORITY
 #endif // ifdef WEBSERVER_HARDWARE
 
 #endif // ifndef WEBSERVER_WEBSERVER_HARDWAREPAGE_H
