+++ src/src/WebServer/Markup_Forms.cpp
@@ -660,6 +660,16 @@ bool getCheckWebserverArg_int(const String& key, int& value) {
   return res;
 }
 
+bool getCheckWebserverArg_int(const String& key,
+                              uint32_t      & value) {
+  const String valueStr = webArg(key);
+  if (valueStr.isEmpty()) return false;
+  uint32_t tmp{};
+  const bool res = validUIntFromString(valueStr, tmp);
+  value = tmp;
+  return res;
+}
+
 bool update_whenset_FormItemInt(const __FlashStringHelper * key,
                                 int         & value) 
 {
@@ -675,6 +685,25 @@ bool update_whenset_FormItemInt(const String& key, int& value) {
   }
   return false;
 }
+
+bool update_whenset_FormItemInt(const __FlashStringHelper * key,
+                                uint32_t    & value) 
+{
+  return update_whenset_FormItemInt(String(key), value);
+}
+
+bool update_whenset_FormItemInt(const String& key,
+                                uint32_t    & value)
+{
+  uint32_t tmpVal;
+
+  if (getCheckWebserverArg_int(key, tmpVal)) {
+    value = tmpVal;
+    return true;
+  }
+  return false;
+}
+
 
 bool update_whenset_FormItemInt(const __FlashStringHelper * key,
                                 int8_t& value)
