+++ src/src/WebServer/Markup_Forms.h
@@ -318,11 +318,20 @@ int getFormItemInt(const String& key, int defaultValue);
 bool getCheckWebserverArg_int(const String& key,
                               int         & value);
 
+bool getCheckWebserverArg_int(const String& key,
+                              uint32_t      & value);
+
 bool update_whenset_FormItemInt(const __FlashStringHelper * key,
                                 int         & value);
 
 bool update_whenset_FormItemInt(const String& key,
                                 int         & value);
+
+bool update_whenset_FormItemInt(const __FlashStringHelper * key,
+                                uint32_t    & value);
+
+bool update_whenset_FormItemInt(const String& key,
+                                uint32_t    & value);
 
 bool update_whenset_FormItemInt(const __FlashStringHelper * key,
                                 int8_t      & value);
